scrapers/requirements.txt
# Python 3.11+ — Sprint 2 scrapers
requests>=2.31.0
beautifulsoup4>=4.12.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
anthropic>=0.40.0
rapidfuzz>=3.0.0
